rl_darts/requirements.txt
absl-py==1.0.0
pytype==2019.10.17
typing==3.7.4.1

ml-collections==0.1.1
xmanager==0.1.4
dm-launchpad[reverb]==0.5.0
gin-config==0.5.0

numpy==1.22.2
tensorflow==2.8.0
tensorflow-probability==0.8.0
dm-sonnet==2.0.0

gym==0.21.0
baselines==0.1.5
procgen=0.10.7
tf-agents==0.3.0
dm-acme==0.4.0
trfl==1.2.0

pyglove
